+++ platform/nxp/lpc17xx/usb_base.c
@@ -56,7 +56,7 @@ static void configPins(struct UsbBase *device,
       config->dm, config->dp, config->connect, config->vbus
   };
 
-  for (uint8_t index = 0; index < ARRAY_SIZE(pinArray); ++index)
+  for (unsigned int index = 0; index < ARRAY_SIZE(pinArray); ++index)
   {
     if (pinArray[index])
     {
